### 2021-12-17 #4002 @whedon
[PRE REVIEW]: PyBioPAX: biological pathway exchange in Python
**Submitting author:** @bgyori (<a href="http://orcid.org/0000-0001-9439-5346">Benjamin M. Gyori</a>)
**Repository:** <a href="https://github.com/indralab/pybiopax" target ="_blank">https://github.com/indralab/pybiopax</a>
**Version:** 0.1.0
