### 2021-12-18 aca21e5
Improve DOIs in paper.md
