+++ paper/paper.bib
@@ -58,6 +58,7 @@
     address = "Online and Punta Cana, Dominican Republic",
     publisher = "Association for Computational Linguistics",
     url = "https://aclanthology.org/2021.emnlp-main.93",
+    doi = {10.18653/v1/2021.emnlp-main.93},
     pages = "1213--1224",
 }
 
@@ -96,7 +97,7 @@
     number = {1},
     pages = {646},
     keywords = {apoptosis, modeling, rule-based, software engineering},
-    doi = {https://doi.org/10.1038/msb.2013.1},
+    doi = {10.1038/msb.2013.1},
     url = {https://www.embopress.org/doi/abs/10.1038/msb.2013.1},
     eprint = {https://www.embopress.org/doi/pdf/10.1038/msb.2013.1},
     year = {2013}
