### 2021-12-21 #4002 @bgyori
Hi @kyleniemeyer, thanks for starting this off! I fixed the DOIs for the references in the paper.
