### 2022-01-02 #4002 @bgyori
Hi @majensen, yes, I looked at the list starting from the bottom and think the following potential reviewers have relevant expertise:
- rvhonorato
- woolfeh
- ayaanhossain

Outside the JOSS list, the following people have relevant expertise:
- leonweber
- nturaga
- dtenenba
- dhimmel
