### 2022-01-24 #4002 @majensen
Hello @rvhonorato - would you be willing to review [this work](#4002) for [JOSS](https://joss.theoj.org)? I would appreciate your expertise here. Please let me by responding to this thread or reaching out to me at maj \-dot\- fortinbras \-at\- gmail \-dot\- com. Thanks for your consideration!
