### 2022-01-24 #4002 @majensen
Hi @dhimmel - would you be available to review [this work](#4002) for [JOSS](https://joss.theoj.org)? Your experience and expertise would be great to have - please let me know on this thread or at maj -dot- fortinbras -at- gmail -dot- com. Thanks for your consideration! --@majensen
