### 2022-01-24 #4002 @cthoyt
Hi @majensen @bgyori the group behind https://github.com/dmis-lab/BERN2 seems to actively be looking for NEN solutions to improve their NER system, so I suggested on their issue tracker that Gilda would be an appropriate solution. Its lead author `@mjeensung` might be appropriate to invite as a reviewer as they may be looking to the software already.
