### 2022-01-24 #4002 @dhimmel
> Hi @dhimmel - would you be available to review this work for JOSS?

Yes
