### 2022-02-01 #4002 @majensen
Hi @leonweber - would you be willing to review [this work](#4002) for our [Journal of Open Source Software](https://joss.theoj.org)? Details about reviewing for us are available at https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. I would appreciate your expertise!
(If you have any questions, please reach out to me at maj -dot- fortinbras -at- gmail -dot- com)
Thanks --@majensen
