### 2022-02-02 #4002 @majensen
Hey there @sap218 - would you have time and inclination to review [this work](#4002) for [JOSS](https://joss.theoj.org)? It converts OWL biological pathway specifications into Python objects and back; I thought this might be up your alley.
Thanks for your consideration! --@majensen
