### 2022-02-06 #4002 @majensen
@wdduncan - you have time to review [this work](#4002) for [JOSS](https://joss.theo4j.org)? I can really use your assistance - thanks!
--@majensen
