### 2022-02-07 #4002 @majensen
@sara-02 - would you be able to review [this work](#4002) for [JOSS](https://joss.theoj.org)? Thanks very much for your consideration! --@majensen
