### 2022-02-07 #4136 @whedon
[REVIEW]: PyBioPAX: biological pathway exchange in Python
**Submitting author:** @bgyori (<a href="http://orcid.org/0000-0001-9439-5346">Benjamin M. Gyori</a>)
**Repository:** https://github.com/indralab/pybiopax
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.1
