### 2022-02-17 indralab/pybiopax#28 @dhimmel
Visualizing BioPaxModels in Python?
I was curious whether there's any ways to visualize a pathway network after loading it as a `pybiopax.biopax.model.BioPaxModel`? Not sure if visualization is in scope for this package, but a visualization example notebook might mike the models more approachable.

_interested as part of https://github.com/openjournals/joss-reviews/issues/4136_
