### 2022-02-17 #4136 @dhimmel
# Daniel Himmelstein's Review

I completed my checklist above. Some general comments: I think it's helpful to have a Python library to interact with BioPax models. The loaders are versatile and convenient. I explored the package briefly in [this notebook](https://colab.research.google.com/drive/1iNbqZPBy02riZCIgshY_Y_sVdzHJRSZN?usp=sharing). I opened https://github.com/indralab/pybiopax/issues/28 about whether visualization is possible.

The [testing coverage](https://github.com/indralab/pybiopax/runs/4571578639?check_suite_focus=true#step:5:43) is impressive. Installation was easy via pip. Dependencies [are light](https://github.com/indralab/pybiopax/blob/aca21e51769f07a121816004dc143e048b9b9764/setup.py#L39): lxml, requests, tqdm.

I'm slightly worried about link rot for the notebook links in the manuscript, since the link to the master branch rather than using a tag or commit to pin a version:

```
https://nbviewer.org/github/indralab/pybiopax/blob/master/notebooks/Explore%20Pathway%20Commons.ipynb
https://nbviewer.org/github/indralab/pybiopax/blob/master/notebooks/Pathway%20Anticorrelation%20Analysis.ipynb
```

Overall a very valuable software contribution to make working with the cryptic BioPAX format a bit easier and more accessible.
