### 2022-02-17 indralab/pybiopax#28 @bgyori
That's a great question! I have an initial implementation of a networkx representation for objects (nodes) and references (edges) between them. However, this visualizes the raw OWL structure of the model which is not necessarily useful for seeing the actual reactions with their inputs and outputs. For that, https://github.com/PathwayCommons/chibe is a great tool that allows loading a BioPAX OWL file for visualization.
