### 2022-02-18 8ce4e13
Update test to work with Reactome version updates
