### 2022-02-28 indralab/pybiopax#30 @leonweber
State of the field section in JOSS paper
The state of the field section seems to be missing from `paper/paper.md`. In this related work such as [this](https://academic.oup.com/bioinformatics/article/36/15/4377/5841664) should be discussed.
