### 2022-02-28 #4136 @leonweber
# Leon Weber's Review

PyBioPAX is a python library to parse and interact with BioPAX files from python. This package closes an important gap in the python bioinformatics ecosystem, as BioPAX is a widespread standard for distributing pathway models and before PyBioPAX parsing such files in python was cumbersome.

I have used the package [in the past](https://github.com/leonweber/extend), there a data loaders for many use cases which are very convenient. The package is extensively tested and installation is easy with few dependencies.

My only concern is that the `State of the field` section seems to be missing from the paper. I have created an issue for this (https://github.com/indralab/pybiopax/issues/30).
