### 2022-02-28 indralab/pybiopax#30 @bgyori
Hello @leonweber, thank you for the feedback! We are not aware of a requirement to have a separate state of field section in JOSS papers (see https://joss.readthedocs.io/en/latest/submitting.html). We mention several related tools in the Statement of Need section and will include additional ones including the one you posted in an update.
