### 2022-02-28 #4136 @bgyori
Thank you @leonweber, we created a new dedicated subsection called State of the field under the Statement of need section where we discuss other commonly used packages for BioPAX. Let us know if anything further is needed. Thank you!
