### 2022-03-05 #4136 @majensen
@bgyori this is a nice tool and appears to be a very complete representation of the underlying spec. The notebooks are detailed and give a user a way in to the practical use of this framework.

Can you please fix the following issues in the references (line numbers from the latest PDF generation above):

l. 107 - capitalize "-Parser" - per original reference
l. 147 - capitalize "Python" - ''
l. 150 - capitalize "R" - ''
l. 152 - author list is borked
l. 156 - capitalize according to original reference

l. 116 and l. 155 - handling of long author list appears not to have worked well here, and inconsistently from elsewhere (e.g., l. 131)

thx
