+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @Article{demir2013using,
-    Title = {Using biological pathway data with paxtools},
+    Title = {{Using biological pathway data with Paxtools}},
     Author = {Demir, Emek and Babur, {\"O}zg{\"u}n and Rodchenkov, Igor and Aksoy, B{\"u}lent Arman and Fukuda, Ken I and Gross, Benjamin and S{\"u}mer, Onur Sel{\c{c}}uk and Bader, Gary D and Sander, Chris},
     Journal = {PLoS computational biology},
     Volume = {9},
@@ -23,8 +23,8 @@
 }
 
 @Article{demir2010biopax,
-    Title = {The BioPAX community standard for pathway data sharing},
-    Author = {Demir, Emek and Cary, Michael P and Paley, Suzanne and Fukuda, Ken and Lemer, Christian and Vastrik, Imre and Wu, Guanming and D'eustachio, Peter and Schaefer, Carl and Luciano, Joanne and others},
+    Title = {{The BioPAX community standard for pathway data sharing}},
+    Author = {Demir, Emek and Cary, Michael P and Paley, Suzanne and Fukuda, Ken and Lemer, Christian and Vastrik, Imre and Wu, Guanming and D'Eustachio, Peter and Schaefer, Carl and Luciano, Joanne and Schacherer, Frank and Martinez-Flores, Irma and Hu, Zhenjun and Jimenez-Jacinto, Veronica and Joshi-Tope, Geeta and Kandasamy, Kumaran and Lopez-Fuentes, Alejandra C and Mi, Huaiyu and Pichler, Elgar and Rodchenkov, Igor and Splendiani, Andrea and Tkachev, Sasha and Zucker, Jeremy and Gopinath, Gopal and Rajasimha, Harsha and Ramakrishnan, Ranjani and Shah, Imran and Syed, Mustafa and Anwar, Nadia and Babur, Özgün and Blinov, Michael and Brauner, Erik and Corwin, Dan and Donaldson, Sylva and Gibbons, Frank and Goldberg, Robert and Hornbeck, Peter and Luna, Augustin and Murray-Rust, Peter and Neumann, Eric and Ruebenacker, Oliver and Samwald, Matthias and van Iersel, Martijn and Wimalaratne, Sarala and Allen, Keith and Braun, Burk and Whirl-Carrillo, Michelle and Cheung, Kei-Hoi and Dahlquist, Kam and Finney, Andrew and Gillespie, Marc and Glass, Elizabeth and Gong, Li and Haw, Robin and Honig, Michael and Hubaut, Olivier and Kane, David and Krupa, Shiva and Kutmon, Martina and Leonard, Julie and Marks, Debbie and Merberg, David and Petri, Victoria and Pico, Alex and Ravenscroft, Dean and Ren, Liya and Shah, Nigam and Sunshine, Margot and Tang, Rebecca and Whaley, Ryan and Letovksy, Stan and Buetow, Kenneth H and Rzhetsky, Andrey and Schachter, Vincent and Sobral, Bruno S and Dogrusoz, Ugur and McWeeney, Shannon and Aladjem, Mirit and Birney, Ewan and Collado-Vides, Julio and Goto, Susumu and Hucka, Michael and Le Novère, Nicolas and Maltsev, Natalia and Pandey, Akhilesh and Thomas, Paul and Wingender, Edgar and Karp, Peter D and Sander, Chris and Bader, Gary D},
     Journal = {Nature biotechnology},
     Volume = {28},
     Number = {9},
@@ -35,9 +35,9 @@
 }
 
 @Article{rodchenkov2020pathway,
-    Title = {Pathway Commons 2019 Update: integration, analysis and exploration of pathway data},
[…]
