### 2022-03-06 #4136 @majensen
Great - thanks @bgyori  - the last item before we ask the associate editors-in-chief to review is to create an open archive of the the latest state of the code repo. This can be done with Zenodo, FigShare, or similar. (Here is a brief tutorial using Zenodo: https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729.)
JOSS prefers that the name of the archive match that of the paper, and that the author list be the same as well.
Once this is accomplished, could you pls put the DOI of the archive here as a comment. I will then make some settings and invoke the AEIC.
Thanks!
