### 2022-03-06 #4136 @majensen
Help @xuanxu - there was a fail in the crossref xml validation (see  https://github.com/openjournals/joss-papers/runs/5440427063?check_suite_focus=true#step:6:53). But here is the crossref element:
`<isbn>1367-4811 (Electronic)</isbn>`
so maybe the validation is too strict?

thanks
