### 2022-03-06 #4136 @xuanxu
@majensen in this case it looks like that entry is incorrect. The ISBN [is invalid](https://isbnsearch.org/search?s=1367-4811) and the value should not include the `(Electronic)` part.
In the bib file the isbn and issn [are the same](https://github.com/indralab/pybiopax/blob/master/paper/paper.bib#L123-L124), I guess that is an error.

And (not related) in that same reference the [url value](https://github.com/indralab/pybiopax/blob/master/paper/paper.bib#L131) looks wrong too.
