### 2022-03-06 #4136 @majensen
Thanks @xuanxu  - since the Crossref XML is created by the bot, not sure what the options would be.
