+++ paper/paper.bib
@@ -120,7 +120,6 @@
 @article{hoyt2017pybel,
     author = {Hoyt, Charles Tapley and Konotopez, Andrej and Ebeling, Christian},
     doi = {10.1093/bioinformatics/btx660},
-    isbn = {1367-4811 (Electronic)},
     issn = {1367-4811},
     journal = {Bioinformatics (Oxford, England)},
     month = {feb},
@@ -128,7 +127,7 @@
     pages = {703--704},
     pmid = {29048466},
     title = {{PyBEL: a computational framework for Biological Expression Language.}},
-    url = {https://www.ncbi.nlm.nih.gov/pubmed/29048466 http://www.ncbi.nlm.nih.gov/pubmed/29048466 http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=PMC5860616},
+    url = {https://www.ncbi.nlm.nih.gov/pubmed/29048466},
     volume = {34},
     year = {2018}
 }
