### 2022-03-09 #4136 @majensen
@bgyori. yes, I'm afraid this is in the hands of other volunteer, unpaid people with day jobs now, the @openjournals/joss-eics.
